+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 3.53.0
 * Add `refund_from_transaction_fee_amount` field to paypal_details
 * Add `refund_from_transaction_fee_currency_iso_code` field to paypal_details
+* Add `revoked_at` field to paypal_account
+* Add support for `PaymentMethodRevokedByCustomer` webhook
 
 ## 3.52.0
 * Deprecate `GrantedPaymentInstrumentUpdate` and add `GrantorUpdatedGrantedPaymentMethod` and `RecipientUpdatedGrantedPaymentMethod`